+++ tests/conftest.py
@@ -29,42 +29,9 @@ def settings_default():
 
 
 @pytest.fixture
-def settings_posix_tmp():
-    """Settings with output dir as '/tmp'"""
-    # TODO: use tempfile module: https://docs.python.org/3/library/tempfile.html#tempfile.gettempdir
-    settings = {
-        "debug": False,
-        "templates_dir": Path(__file__).parent.with_name("faddr").joinpath("templates"),
-        "database": {"path": "/tmp/", "name": "faddr-db.sqlite"},
-        "rancid": {
-            "dirs": [
-                {
-                    "path": "tests/fixtures/rancid/",
-                    "kind": "dir",
-                    "mapping": {"cisco-faddr": "cisco-ios"},
-                }
-            ],
-            "default_mapping": {
-                "cisco": "cisco-ios",
-                "cisco-xr": "cisco-iosxr",
-                "juniper": "juniper-junos",
-            },
-            "mapping": {},
-        },
-    }
-    return settings
-
-
-@pytest.fixture
 def settings_file_absent():
     """Absent settings file."""
     return Path("tests/fixtures/nonexistant_settings_file.yaml")
-
-
-@pytest.fixture
-def settings_file_posix_tmp():
-    """Settings file with output dir as '/tmp'."""
-    return Path("tests/fixtures/faddr_posix_tmp.yaml")
 
 
 @pytest.fixture